heap/non-heap memory as a percentage of committed
Memory utilization (heap and non-heap) has been calculated as a percentage of
the maximum, this was lifted from the Dropwizard Metrics code this project
replaced.  However, on our JDK 8 installations, `max' for NonHeapUsage is
undefined (-1), resulting in obviously wrong values.

Committed seems a better divisor here because a) according the contract it will
always be defined, and b) of the two it represents the only guarantee of how
much memory is available.

Bug: https://phabricator.wikimedia.org/T101764