daemonization
Starts as a "service" (stays in the foreground and logs to standard out). Accepts
flags specifying the collection interval, and carbon host/port (run with
{-h,--help} for a synopsis).

Uses the new Daemon class as the main for the jar files, (the old Command class
is still there, but must be fully qualified, see the README).  Since this is
breaking, the version major has been bumped to 2.

The dependency on the JDK persists, but munging of the classpath has been moved
into the Java code, eliminating the wrapper script.  Hopefully this proves less
brittle.

Bug: https://phabricator.wikimedia.org/T95253