Port to Cassandra 2.2
Somewhere between 2.1.13, and 2.2.5, Cassandra was refactored to wrap all of
the codahale/dropwizard metric mbeans in Cassandra delegators.  This would
have been awesome if it were done this way from the beginning, but
after-the-fact it broke cmcd by changing the object names of everything.

This is an alternative to what was proposed in
wikimedia/cassandra-metrics-collector/pull/7.  Instead of supporting both
versions with a JmxCollectorFactory, this changeset eschews 2.1 compat, (and
is simpler as a result).

Bug: https://phabricator.wikimedia.org/T126629