timeout long running jobs
There is evidence to suggest that cmcd jobs are blocking for long periods of
time while writing to Carbon (presumably because of issues with Carbon).  When
this happens, the jobs begin piling up locally.

At a minimum, it seems unwise to allow issues with the remote-side to create an
unbounded number of in-flight jobs.  Additionally, if the remote is struggling,
it is likely counter-productive to ratchet up the concurrency in an attempt at
storing every last metric.

This changeset introduces a TimedTask object which is used to run collections,
raising an exception and terminating the job if it runs for longer than the
configured collection interval, or 60 seconds, whichever is smaller.